backend/pyproject.toml
[tool.poetry]
name = "pro-web-app"
version = "0.1.0"
description = "Professional FastAPI app with Poetry, PostgreSQL and Docker"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.12"

fastapi = "^0.115.0"
uvicorn = { extras = ["standard"], version = "^0.32.0" }

sqlalchemy = "^2.0.36"
asyncpg = "^0.30.0"
alembic = "^1.14.0"

pydantic = "^2.10.0"
pydantic-settings = "^2.6.1"
python-multipart = "^0.0.22"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
bcrypt = "4.0.1"
python-jose = {extras = ["cryptography"], version = "^3.5.0"}

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.24.0"
httpx = "^0.28.0"
aiosqlite = "^0.20.0"

ruff = "^0.8.0"
pre-commit = "^4.0.1"
pytest-cov = "^7.1.0"

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["B008"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]